requirements.txt
flask==2.1.3
gunicorn==19.9.0
requests-toolbelt==0.9.1
requests==2.28.1
keras==2.9.0
tensorflow-cpu==2.9.0
numpy==1.22.3
pandas==1.4.3